bountyhound/requirements/requirements-cloud.txt
# Cloud Infrastructure Testing Requirements

# AWS
boto3==1.35.80
botocore==1.35.80
awscli==1.36.5

# Azure
azure-cli==2.66.0
azure-identity==1.19.0
azure-mgmt-storage==21.2.1

# GCP
google-cloud-storage==2.18.2
google-auth==2.36.0
google-cloud-iam-credentials==1.0.0
google-cloud-iam==2.15.2
google-cloud-resource-manager==1.12.5
google-cloud-functions==1.16.5
google-cloud-firestore==2.18.0
google-cloud-secret-manager==2.20.2

# Cloud Security Tools
ScoutSuite==5.14.0

# Utilities
requests==2.32.3
colorama==0.4.6
